refactor(process): migrate hostinfo component to v2
