+++ comp/process/hostinfo/def/component.go
