+++ comp/process/hostinfo/impl/hostinfo.go
@@ -9,25 +9,17 @@ package hostinfoimpl
 import (
 	"fmt"
 
-	"go.uber.org/fx"
-
 	"github.com/DataDog/datadog-agent/comp/core/config"
 	"github.com/DataDog/datadog-agent/comp/core/hostname/hostnameinterface"
 	ipc "github.com/DataDog/datadog-agent/comp/core/ipc/def"
 	log "github.com/DataDog/datadog-agent/comp/core/log/def"
-	hostinfoComp "github.com/DataDog/datadog-agent/comp/process/hostinfo"
+	compdef "github.com/DataDog/datadog-agent/comp/def"
+	hostinfoComp "github.com/DataDog/datadog-agent/comp/process/hostinfo/def"
 	"github.com/DataDog/datadog-agent/pkg/process/checks"
-	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
 
-// Module defines the fx options for this component.
-func Module() fxutil.Module {
-	return fxutil.Component(
-		fx.Provide(newHostInfo))
-}
-
 type dependencies struct {
-	fx.In
+	compdef.In
 
 	Config   config.Component
 	Hostname hostnameinterface.Component
@@ -39,7 +31,8 @@ type hostinfo struct {
 	hostinfo *checks.HostInfo
 }
 
-func newHostInfo(deps dependencies) (hostinfoComp.Component, error) {
+// NewComponent creates a new hostinfo component.
+func NewComponent(deps dependencies) (hostinfoComp.Component, error) {
 	hinfo, err := checks.CollectHostInfo(deps.Config, deps.Hostname, deps.IPC)
 	if err != nil {
 		_ = deps.Logger.Critical("Error collecting host details:", err)
